exit checker package and its test
